requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1
django-allauth==0.54.0
django-cors-headers==4.1.0
django-import-export==3.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
MarkupPy==1.14
numpy==1.25.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.2
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tablib==3.5.0
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0